moves view code out of previewViewController into own class, basic set up done. todo: animating keyboard, adjusting textview height
